Finished adding the passive item upgrade system. Cleaned the code Items, Options, PassiveSlot and Slot.
Finished adding the passive item upgrade system, still have to correct a bug in which something not wanted display on the upgrade screen.
Cleaned the code of Items, Options, PassiveSlot and Slot. Notably added summary, managed the setter and getter to only keep the necessary ones, removed some useless redundant bits of code.