Finished the whole architecture of the items class.
Now each item have the same Items.cs parent class, then Passives.cs or Weapons.cs class in function of their type. The Items parent class manage the effect updates and the levels and upgrades system.

I'll continue to make some tweaks on the class to finish it and to make it even more practical.

I still have to implement it in the upgrade menu.