Changed the structure of the item. Added global class for global access to some data. Updated the upgreade system with the new item structure.
Item is now composed of one Item class that manage all the generic logic of items that can be attached to a resource that add the custom logic for each item.

Global access to the player reference and all possible passive and weapons.

Started to work on the Item display system. Still have to finish the upgrade system so that a player can choose new passive along the way.